packages/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ai-query",
    "aiohttp-cors>=0.8.1",
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.1",
    "snscrape>=0.7.0.20230622",
    "tavily-python>=0.7.17",
]

[tool.uv.sources]
ai-query = { path = "../../../../ai-query", editable = true }
snscrape = { git = "https://github.com/JustAnotherArchivist/snscrape.git" }
